pyproject.toml
[project]
name = "chatbot-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cdp-langchain>=0.0.13",
    "langchain-anthropic>=0.3.7",
    "langchain-core>=0.3.33",
    "langchain-openai==0.2.9",
    "langgraph>=0.2.69",
    "matplotlib>=3.10.0",
    "mplfinance>=0.12.10b0",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
]
